refactor: remove reactn and old store refactor: rename "store_new" to "store"
