+++ package.json
@@ -54,7 +54,6 @@
     "react-scripts": "3.1.2",
     "react-tagsinput": "^3.19.0",
     "react-toastify": "^5.4.0",
-    "reactn": "^2.2.2",
     "redux": "^4.0.4",
     "redux-persist": "^6.0.0",
     "semantic-ui-css": "^2.4.1",
