+++ src/components/Searchbar.tsx
@@ -1,4 +1,4 @@
-import React from 'reactn';
+import React from 'react';
 
 export type SearchType = 'video' | 'segment';
