+++ src/components/YouTubePlayer.tsx
@@ -1,7 +1,7 @@
 /* global YT */
 // https://developers.google.com/youtube/iframe_api_reference
 // https://developers.google.com/youtube/player_parameters
-import React from 'reactn';
+import React from 'react';
 import nullthrows from 'nullthrows';
 
 export type YouTubePlayerState =
